refactor[#345] : 마킹 추가 api 요청 함수에 apiClient 적용
